Remove ParticleAttributeSize and openPMDCountParticles
Legacy from ADIOS Writer
